+++ dbuzzy/Cargo.toml
@@ -1,17 +1,22 @@
 [package]
-name = "duzzy-db"
+name = "duzzy"
 version.workspace = true
 edition.workspace = true
 authors.workspace = true
 
 [dependencies]
 anyhow.workspace = true
+thiserror.workspace = true
 tokio.workspace = true
+ratatui.workspace = true
+crossterm.workspace = true
 
-thiserror = "1.0.61"
+futures-util = "0.3.30"
+
+tokio-postgres = "0.7.10"
+deadpool-postgres = "0.14.0"
 
 toml = "0.8.14"
 serde = { version = "1.0.203", features = ["derive"] }
 
-tokio-postgres = "0.7.10"
-deadpool-postgres = "0.14.0"
+duzzy-editor = { path = "../duzzy-editor" }
