+++ dbuzzy/src/db/mod.rs
@@ -1,11 +1,5 @@
-#![warn(
-    clippy::perf,
-    clippy::semicolon_if_nothing_returned,
-    clippy::missing_const_for_fn,
-    clippy::use_self
-)]
-pub mod config;
-pub mod connection;
+mod config;
+mod connection;
 
 pub type DbResult<T> = anyhow::Result<T, Error>;
 
@@ -13,6 +7,8 @@ pub type DbResult<T> = anyhow::Result<T, Error>;
 pub enum Error {
     #[error("Pool error: {0}")]
     PoolError(#[from] deadpool_postgres::PoolError),
+    #[error("Create Pool error: {0}")]
+    CreatePoolError(#[from] deadpool_postgres::CreatePoolError),
     #[error("PostgreSQL error: {0}")]
     PgError(#[from] tokio_postgres::Error),
 }
