+++ dbuzzy/src/main.rs
@@ -1,6 +1,13 @@
+#![warn(
+    clippy::perf,
+    clippy::semicolon_if_nothing_returned,
+    clippy::missing_const_for_fn,
+    clippy::use_self
+)]
 use ratatui::backend::CrosstermBackend;
 
 mod app;
+mod db;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
