--- duzzy/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "duzzy"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-
-[dependencies]
-log.workspace = true
-anyhow.workspace = true
-tokio.workspace = true
-ratatui.workspace = true
-crossterm.workspace = true
-
-futures-util = "0.3.30"
-
-duzzy-editor = { path = "../duzzy-editor" }
